refactor(web): split api.test.ts into per-module colocated test files
Split the monolithic 2632-line api.test.ts into focused per-module files:
- event-parser.test.ts (64 tests)
- stream.test.ts (23 tests — polling reconciliation + idle watchdog)
- messages.test.ts (15 tests — normalizeContentOrder/TextSegments + wire format)
- post-chat-message.test.ts (5 tests — onboarding payload via fetch mock)
- conversations.test.ts (7 tests — parseConversation originChannel)
- diagnostics-platform.test.ts (3 tests — recordChatDiagnostic platform tag)

Merges normalizeContentOrder/TextSegments tests into the existing
messages.test.ts. Moves recordChatDiagnostic platform tests to
utils/diagnostics-platform.test.ts alongside the existing diagnostics.test.ts.

Total test count preserved: 117 tests across 6 files.

Closes LUM-1725
